+++ CHANGES.md
@@ -6,6 +6,9 @@ This release supports [version
 
 ## Changes
 
+* We have a new command `mir_find_name : MIRModule -> String -> [MIRType] -> String`
+  which can be used to find the name of the instantiation of a polymorphic function.
+
 * The REPL's `:search` command can now handle unapplied type constructors. For
   example, one can now write `:search ProofScript` to find objects mentioning
   `ProofScript` applied to any type.